refactor InfiniteScroll implementation
- created a wrapper component to simplify interface
- rearranged code
