+++ src/types/helpers.d.ts
@@ -1,0 +1,1 @@
+export type Overwrite<T, U> = Pick<T, Exclude<keyof T, keyof U>> & U
